Improve global search dialog
