+++ frontend/src/components/AppIcon.css
@@ -1,0 +1,27 @@
+/* Shared artwork/fallback behavior; each surface owns only icon size and radius. */
+.app-icon {
+  position: relative;
+  flex-shrink: 0;
+  display: grid;
+  place-items: center;
+  overflow: hidden;
+  background: var(--app-color, var(--accent));
+  color: var(--accent-fg, #fff);
+  font-weight: 750;
+}
+
+.app-icon img {
+  position: absolute;
+  inset: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: contain;
+}
+
+.app-icon.is-image {
+  overflow: visible;
+  border-radius: 0;
+  background: transparent;
+}
+
+.app-icon.is-image > span { visibility: hidden; }
