+++ frontend/src/components/Drawer/drawerInformationArchitecture.js
@@ -89,14 +89,3 @@ export function filterInstalledApps(apps = [], query = '') {
       .includes(needle)
   ))
 }
-
-export function appInitials(name) {
-  const words = String(name || '')
-    .replace(/[^a-z0-9]+/gi, ' ')
-    .trim()
-    .split(/\s+/)
-    .filter(Boolean)
-  if (words.length === 0) return 'A'
-  if (words.length === 1) return words[0].slice(0, 2).toLocaleUpperCase()
-  return `${words[0][0]}${words[1][0]}`.toLocaleUpperCase()
-}
